requirements.txt
ujson
motor
sanic
aiohttp
passlib
marshmallow
python-whois
cached-property
